llma/ws/one/requirements.txt
requests>=2.30
openpyxl>=3.1.0
python-dotenv
colorama